+++ .gitignore
@@ -1,1 +1,5 @@
 build
+.vscode
+*.code-workspace
+install
+log
