+++ slam_vdb/cmake/slam_vdb-config.cmake
@@ -1,0 +1,6 @@
+# This file is part of slam_vdb
+# Sriharsha Ghanta <ghanta1996@gmail.com>
+
+# TODO : Add the license here
+
+include("${CMAKE_CURRENT_LIST_DIR}/slam_vdb-targets.cmake")
