+++ slam_vdb/include/slam_vdb/measurement/point/point_base.hpp
