+++ slam_vdb/include/slam_vdb/measurement/point/point_xyz.hpp
