+++ slam_vdb/include/slam_vdb/measurement/pointcloud/pointcloud_base.hpp
