+++ slam_vdb/include/slam_vdb/slam_vdb.hpp
@@ -4,4 +4,14 @@
 #include "slam_vdb/measurement/point/point_xyz.hpp"
 #include "slam_vdb/measurement/pointcloud/pointcloud.hpp"
 
+namespace slam_vdb {
+class slam_vdb
+{
+private:
+  /* data */
+public:
+  slam_vdb(/* args */);
+  ~slam_vdb();
+};
+}  // namespace slam_vdb
 #endif  // SLAM_VDB_SLAM_VDB_HPP_
