+++ slam_vdb/src/slam_vdb.cpp
@@ -1,0 +1,7 @@
+#include "slam_vdb/slam_vdb.hpp"
+
+namespace slam_vdb {
+slam_vdb::slam_vdb() {}
+
+slam_vdb::~slam_vdb() {}
+}  // namespace slam_vdb
