+++ slam_vdb/tests/measurement/point/test_point_xyz.cpp
