+++ slam_vdb/tests/measurement/pointcloud/test_pointcloud.cpp
