+++ slam_vdb/tests/test_main.cpp
